requirements.txt
numpy
pandas
ipython
ipykernel
# Trick to install the version of Tensorflow depending on your processor: darwin == Mac, ARM == M1
tensorflow-macos==2.15.0; sys_platform == 'darwin' and 'ARM' in platform_version # Mac M-chips
tensorflow==2.15.0; sys_platform == 'darwin' and 'ARM' not in platform_version # Mac Intel chips
tensorflow==2.15.0; sys_platform != 'darwin' # Windows & Ubuntu Intel chips
uvicorn
fastapi
Pillow
matplotlib
python-multipart
opencv-python